tuxun_agent/requirements.txt
fastapi==0.104.1  
uvicorn==0.24.0  
python-multipart==0.0.6  
opencv-python==4.8.1.78  
pillow==10.1.0  
exifread==3.0.0  
numpy==1.24.3  
torch==2.1.1  
torchvision==0.16.1  
faiss-cpu==1.7.4  
openai==1.3.5  
requests==2.31.0  
geopy==2.4.0 
